fix prototype
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@31301 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/unix/ps-file.c
@@ -199,7 +199,7 @@ void ps_file_out_data(BYTE outval, int chipno)
     }
 }
 
-void ps_file_out_ctr(WORD parsid_ctrport, int chipno)
+void ps_file_out_ctr(BYTE parsid_ctrport, int chipno)
 {
     BYTE ctl = 0;
     int datadir = 0;
